Replace Dist::Zilla plugin
...with @SCHWIGON.

Dist::Milla is way too opinionated. It has no way to
override its opinions. That makes it especially hard
when the (co-)maintainer changes.

So we replace it with another plugin which is similarly
complex but allows overwriting. A small downside is the
need to provide some more basic facts.